refactor(web-client): factor out common handleScan helper
