+++ web-client/src/app/views/wallet-access/wallet-access.page.ts
@@ -7,9 +7,8 @@ import { ScannerService } from 'src/app/services/scanner.service';
 import { WalletService } from 'src/app/services/wallet/wallet.service';
 import { SessionQuery } from 'src/app/stores/session/session.query';
 import { SwalHelper } from 'src/app/utils/notification/swal-helper';
-import { never } from 'src/helpers/helpers';
 import { LockscreenPage } from '../lockscreen/lockscreen.page';
-import { ScannerPage, ScanResult } from '../scanner/scanner.page';
+import { handleScan } from '../scanner.helpers';
 
 @Component({
   selector: 'app-wallet-access',
@@ -45,59 +44,8 @@ export class WalletAccessPage implements OnInit {
     }
   }
 
-  // FIXME: Duplication with SendFundsPage.presentScanner
   async openScanner() {
-    const scanSuccess = async (address: string) => {
-      this.address = address;
-      await this.confirm();
-    };
-
-    // FIXME: fix import for OverlayEventDetail
-    const dismissed = async (eventDetail: { data?: ScanResult }) => {
-      const { data: result } = eventDetail;
-      switch (result?.type) {
-        case 'scanSuccess':
-          await scanSuccess(result.result);
-          break;
-        case 'scanError':
-          await this.notification.swal.fire({
-            icon: 'error',
-            title: 'Error scanning QR code',
-            text: 'Failed to scan a valid QR code. Please try again.',
-          });
-          break;
-        case 'camerasNotFound':
-          await this.notification.swal.fire({
-            icon: 'warning',
-            title: 'No camera found',
-            text: 'In order to scan a QR Code, your device must have a camera',
-          });
-          break;
-        case 'permissionDenied':
-          await this.notification.swal.fire({
-            icon: 'error',
-            title: 'Permission required',
-            text: `In order to scan a QR Code, you need to grant camera's permission`,
-          });
-          break;
-        case 'dismissed':
-          // Explicit user dismiss: just return silently.
-          break;
-        case undefined:
-          throw Error(
-            'ScannerPage modal dismiss returned unexpected undefined!'
-          );
-        default:
-          never(result);
-      }
-    };
-
-    // Show modal
-    const scanner = await this.modalCtrl.create({ component: ScannerPage });
-    const didDismissPromise = scanner.onDidDismiss();
-
-    await scanner.present();
-    await dismissed(await didDismissPromise);
+    await handleScan(this.modalCtrl, this.notification.swal, this.confirm);
   }
 
   async confirm() {
